requirements.txt
psycopg2
python-dotenv
mysql-connector-python